refactor: change password option location
