+++ frontend/src/shared/i18n/en-GB.json
@@ -308,6 +308,7 @@
         }
       },
       "formerMembers": "Former members",
+      "hideFormerMembers": "Hide old members",
       "infoLine": {
         "events": "event(s)",
         "members": "member(s)"
@@ -350,6 +351,7 @@
       "requestAdmin": "Ask to be an admin",
       "requestSent": "Request sent",
       "requestToBeAdmin": "{{student}} asks to be admin",
+      "showFormerMembers": "Show old members",
       "subGroupOf": "Subgroup of",
       "subgroups": "Subgroups",
       "tabs": {
@@ -758,11 +760,36 @@
   },
   "student": {
     "arrivalYear": "Arrival year",
+    "confirmNewPassword": "Confirm new password",
+    "currentPassword": {
+      "label": "Current password"
+    },
+    "description": {
+      "label": "Description",
+      "placeholder": "Describe yourself in one or two sentences"
+    },
     "details": {
-      "changePassword": "Change password"
+      "changePassword": "Change password",
+      "editPassword": "Change password",
+      "editProfile": "Edit your profile",
+      "form": {
+        "firstName": {
+          "label": "First name"
+        }
+      },
+      "housing": "Shared accommodation",
+      "links": "Links",
+      "password": "Password",
+      "profile": "Profile"
     },
     "filters": {
       "title": "Filters"
+    },
+    "firstName": {
+      "label": "First name"
+    },
+    "lastName": {
+      "label": "Last Name"
     },
     "list": {
       "labelDisplayedRows": "Page {{ page }} of {{ pageSize }} ({{ count }} results)"
@@ -770,12 +797,18 @@
     "name": {
       "label": "Name"
     },
+    "newPassword": {
+      "label": "New Password"
+    },
     "picture": {
       "helperText": "Recommended size: 400 x 400",
       "label": "Picture"
     },
     "searchBar": {
       "placeholder": "Search for a student"
+    },
+    "username": {
+      "label": "User name"
     }
   },
   "userMenu": {
